+++ rust/segment/src/lib.rs
@@ -1,8 +1,8 @@
+pub mod blockfile_metadata;
+pub mod blockfile_record;
 pub mod config;
-pub mod distributed_hnsw_segment;
-pub mod metadata_segment;
-pub mod record_segment;
-pub mod spann_segment;
+pub mod distributed_hnsw;
+pub mod distributed_spann;
 pub mod test;
 pub mod types;
 pub mod utils;
