+++ rust/worker/src/execution/operators/knn_hnsw.rs
@@ -4,7 +4,7 @@ use chroma_error::{ChromaError, ErrorCodes};
 use chroma_types::SignedRoaringBitmap;
 use thiserror::Error;
 
-use chroma_segment::distributed_hnsw_segment::DistributedHNSWSegmentReader;
+use chroma_segment::distributed_hnsw::DistributedHNSWSegmentReader;
 use chroma_system::Operator;
 
 use super::knn::{KnnOperator, RecordDistance};
