+++ rust/worker/src/execution/operators/spann_centers_search.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use chroma_distance::DistanceFunction;
 use chroma_error::{ChromaError, ErrorCodes};
 use chroma_index::spann::utils::rng_query;
-use chroma_segment::spann_segment::{SpannSegmentReader, SpannSegmentReaderContext};
+use chroma_segment::distributed_spann::{SpannSegmentReader, SpannSegmentReaderContext};
 use chroma_system::Operator;
 use thiserror::Error;
