+++ rust/worker/src/execution/orchestration/compact.rs
@@ -31,11 +31,11 @@ use chroma_blockstore::provider::BlockfileProvider;
 use chroma_error::ChromaError;
 use chroma_error::ErrorCodes;
 use chroma_index::hnsw_provider::HnswIndexProvider;
-use chroma_segment::distributed_hnsw_segment::DistributedHNSWSegmentWriter;
-use chroma_segment::metadata_segment::MetadataSegmentWriter;
-use chroma_segment::record_segment::RecordSegmentReader;
-use chroma_segment::record_segment::RecordSegmentReaderCreationError;
-use chroma_segment::record_segment::RecordSegmentWriter;
+use chroma_segment::distributed_hnsw::DistributedHNSWSegmentWriter;
+use chroma_segment::blockfile_metadata::MetadataSegmentWriter;
+use chroma_segment::blockfile_record::RecordSegmentReader;
+use chroma_segment::blockfile_record::RecordSegmentReaderCreationError;
+use chroma_segment::blockfile_record::RecordSegmentWriter;
 use chroma_segment::types::ChromaSegmentFlusher;
 use chroma_segment::types::ChromaSegmentWriter;
 use chroma_segment::types::MaterializeLogsResult;
